# 003-APIRESTOpenAPI/MiClienteReservas/api/openapi.yaml
openapi: 3.0.4
info:
  title: DemoApi
  version: "1.0"
servers:
- url: /
security:
- Bearer: []
...
paths:
  /api/auth/login:
    post:
      operationId: Login
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: "#/components/schemas/LoginRequestDto"
              description: Credenciales utilizadas para solicitar un token de acceso.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectApiResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectApiResponse"
          description: Bad Request
        "401":
          description: Unauthorized
      tags:
      - Auth
  /api/categories:
    get:
      operationId: GetCategories
      responses:
        "401":
          description: Unauthorized
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryListApiResponse"
          description: OK
      tags:
      - Category
    post:
      operationId: CreateCategory
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: "#/components/schemas/CreateCategoryRequestDto"
              description: Datos necesarios para crear una categoría.
        required: true
      responses:
        "401":
          description: Unauthorized
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiResponse"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiResponse"
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiResponse"
          description: Conflict
      tags:
      - Category
  /api/categories/{id}:
    delete:
      operationId: DeleteCategory
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
        style: simple
      responses:
        "401":
          description: Unauthorized
        "204":
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiResponse"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiResponse"
          description: Conflict
      tags:
      - Category
    get:
      operationId: GetCategoryById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
        style: simple
      responses:
        "401":
          description: Unauthorized
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiResponse"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiResponse"
          description: Not Found
      tags:
      - Category
    put:
      operationId: UpdateCategory
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: "#/components/schemas/UpdateCategoryRequestDto"
              description: Datos que pueden modificarse de una categoría existente.
        required: true
      responses:
        "401":
          description: Unauthorized
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiResponse"
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiResponse"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiResponse"
          description: Conflict
      tags:
      - Category
...
components:
  schemas:
    Category:
      additionalProperties: false
      description: Representa una categoría para clasificar productos.
      example:
        id: 0
        name: name
        description: description
      properties:
        id:
          description: Identificador único de la categoría.
          format: int32
          type: integer
        name:
          description: Nombre visible de la categoría.
          nullable: true
          type: string
        description:
          description: Descripción de la categoría.
          nullable: true
          type: string
      type: object
    CategoryApiResponse:
      additionalProperties: false
      example:
        isSuccess: true
        message: message
        data: ""
        errors:
        - errors
        - errors
      properties:
        isSuccess:
          type: boolean
        message:
          nullable: true
          type: string
        data:
          allOf:
          - $ref: "#/components/schemas/Category"
          description: Representa una categoría para clasificar productos.
          nullable: true
        errors:
          items:
            type: string
          nullable: true
          type: array
      type: object
    CategoryListApiResponse:
      additionalProperties: false
      example:
        isSuccess: true
        message: message
        data:
        - id: 0
          name: name
          description: description
        - id: 0
          name: name
          description: description
        errors:
        - errors
        - errors
      properties:
        isSuccess:
          type: boolean
        message:
          nullable: true
          type: string
        data:
          items:
            $ref: "#/components/schemas/Category"
          nullable: true
          type: array
        errors:
          items:
            type: string
          nullable: true
          type: array
      type: object
    CreateCategoryRequestDto:
      additionalProperties: false
      description: Datos necesarios para crear una categoría.
      example:
        name: name
        description: description
      properties:
        name:
          description: Nombre visible de la categoría.
          maxLength: 100
          minLength: 2
          type: string
        description:
          description: Descripción de la categoría.
          maxLength: 500
          minLength: 0
          nullable: true
          type: string
      required:
      - name
      type: object
...
    LoginRequestDto:
      additionalProperties: false
      description: Credenciales utilizadas para solicitar un token de acceso.
      example:
        username: username
        password: password
      properties:
        username:
          description: Nombre del usuario.
          minLength: 1
          type: string
        password:
          description: Contraseña del usuario.
          minLength: 1
          type: string
      required:
      - password
      - username
      type: object
    ObjectApiResponse:
      additionalProperties: false
      example:
        isSuccess: true
        message: message
        data: ""
        errors:
        - errors
        - errors
      properties:
        isSuccess:
          type: boolean
        message:
          nullable: true
          type: string
        data:
          nullable: true
        errors:
          items:
            type: string
          nullable: true
          type: array
      type: object
...
    UpdateCategoryRequestDto:
      additionalProperties: false
      description: Datos que pueden modificarse de una categoría existente.
      example:
        name: name
        description: description
      properties:
        name:
          description: Nuevo nombre de la categoría.
          maxLength: 100
          minLength: 2
          type: string
        description:
          description: Nueva descripción de la categoría.
          maxLength: 500
          minLength: 0
          nullable: true
          type: string
      required:
      - name
      type: object
...
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      description: "Escribe: Bearer {token}"
      scheme: bearer
      type: http